% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{collapse_group}
\alias{collapse_group}
\title{looply collapse each vector of a list by a group of indexs}
\usage{
collapse_group(x, g, sep = "", fill = "")
}
\arguments{
\item{x}{list}

\item{g}{list}

\item{sep}{string}

\item{fill}{string}
}
\value{
list
}
\description{
looply collapse each vector of a list by a group of indexs
}

