% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{open_file}
\alias{open_file}
\title{open a vcf file}
\usage{
open_file(path)
}
\arguments{
\item{path}{the path for vcf file}
}
\value{
a reader object for vcf reading
}
\description{
open a vcf file
}

