% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{open_vcf}
\alias{open_vcf}
\title{open a vcf file (gz-compressed or plain)}
\usage{
open_vcf(file)
}
\arguments{
\item{file}{filename}
}
\value{
an vcf object which is an environment including the file connection, meta lines, header, etc.
}
\description{
open a vcf file (gz-compressed or plain)
}

