% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{read_body}
\alias{read_body}
\title{read n variants}
\usage{
read_body(vcf, n = 0)
}
\arguments{
\item{vcf}{a vcf object returned by open_vcf.}

\item{n}{n variants/lines to read}
}
\value{
data.table the vcf table format.
}
\description{
read n variants
}

