% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_lines}
\alias{read_lines}
\title{read n lines from the file.}
\usage{
read_lines(reader, n)
}
\arguments{
\item{reader}{a pointer returned by open_vcf.}

\item{n}{how many lines to read default: n=0 for file end}
}
\value{
a character vector, return a empty vector (length=0) when reading a file that had reached EOF; return a vector of length 1 with "" for blank line.
}
\description{
read n lines from the file.
}

