% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{reformat_body}
\alias{reformat_body}
\title{reformat INFO and FORMAT as varid, key, value tables}
\usage{
reformat_body(lines, varid_offset = 0)
}
\arguments{
\item{lines}{data.table returned by read_body}
}
\value{
list a list of data.tables, including C7: cols1-7, INFO:, FORMAT: FORMAT and samples' genotype information.
}
\description{
reformat INFO and FORMAT as varid, key, value tables
}

