% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{tobody}
\alias{tobody}
\title{revert variants back to the vcf body table format.}
\usage{
tobody(variants)
}
\arguments{
\item{variants}{an object returned by reformat_body.}
}
\value{
data.table
}
\description{
revert variants back to the vcf body table format.
}

